Improve component state management and consistency (#1174)
* Rename query and loader to reflect profile context

Updated query and loader names from COMPONENTS_INDEX_QUERY to PROFILE_INDEX_QUERY to better align with the context of the profile route. This change avoids duplication of query names in code.

* Update component state field to use string with predefined options

Replaced the boolean `documentationIsReady` field with a `componentState` string field. Added predefined options (`In progress`, `Ready`, `Deprecated`) and set `In progress` as the default value. This change enhances flexibility and better represents the component's lifecycle.

* Add Sanity CLI configuration file

Introduce a new CLI configuration file for the Sanity CMS. This file specifies the project ID and dataset to be used, enabling proper setup for CLI operations.

* Refactor Sanity types and queries for improved consistency.

Simplified type definitions by consolidating union syntax and formatting. Added a new "componentState" field along with its possible values and adjusted the schema to align with these changes. Removed unused "documentationIsReady" field to clean up the types.

* Add `state` field to component query and update badge display

Included the `state` field in the GROQ query with a default value of "In progress" and replaced conditional badge rendering with dynamic `state` display. This simplifies the handling of component statuses and makes it easier to add new component states later.

* Update component schema and queries to include documentation state

Added a new "documentationState" field to the component schema with predefined states. Updated all related TypeScript types, GROQ queries, and UI to reflect this change. Adjusted componentState and documentationState options to use consistent casing.

* combine documentState, componentState and highlightAsNew

* remove unused sanity query

* Format

---------

Co-authored-by: Alexander Bjerkan <alexander.bjerkan@obos.no>
Co-authored-by: Oscar Carlström <oscar.carlstrom@gmail.com>