Add layout gap utils (#1415)
## Layout gap utils

Add new utility classes for layout x/y gap. These can be used inside
layout grids/flex containers for consistent layout spacing. This will
also be added to horizontal image-cards in the `react`-package after
this is merged.

Next steps in the tailwind-package will be to add utility-classes for
`layout-grid` so that we can move away from the current situation where
this is duplicated in each team's code base.

I think it makes sense to create a single release for this, so let's
wait until remaning PR's have been approved and merged.