Table (#1437)
This pull request introduces a new, accessible table component system to
the React package, providing a set of composable table primitives with
built-in horizontal scrolling, keyboard navigation, and WCAG compliance.
The changes include the implementation of the table components, their
export from the package, and comprehensive Storybook stories
demonstrating usage and features.

We use the same composition as in other grunnmuren-components, where the
user "builds their own table", and the children are the cells.

Uses the same scrolling and responsiveness as Tabs!



https://github.com/user-attachments/assets/4155d51d-6497-465d-9d73-c35da828e0dc

---------

Co-authored-by: Aulon Mujaj <4094284+aulonm@users.noreply.github.com>