Fix scroll issues in the Carousel component (#1480)
Address scrolling bugs in the Carousel component by implementing a queue
for scroll actions and debouncing scroll events to enhance user
experience.

Also disabling the default scroll staggering when user is holding down
the arrow keys which is the undesired native behavior of
scroll-snapping.

Still not 100%, but far better than before.

Before:


After:

https://github.com/user-attachments/assets/1e33ef3a-63c0-406d-8321-44ab27d44f60

---------

Co-authored-by: Aulon Mujaj <4094284+aulonm@users.noreply.github.com>