Extract LinkList styles and expose LinkListContainer (#1521)
Extracting styles for the `<LinkList>` components into reusable classes.
This change enhances the readability of the implementation. And allows
the classes to be used outisde React.

Exposing `<LinkListContainer>` as part of the `<LinkList>` API. This
allows for easier customization and flexibility. Since it is now
possible to style the container and the list individually. This means
you can still just render shorter lists (less than 6 LinkListItems) like
before.

This also paves way for supporting `<Heading>` inside the
`<LinkListContainer>`, above the `<LinkListContainer>`. This will be
done in the next PR!



https://github.com/user-attachments/assets/a6d8bb58-2ba5-41e7-82cf-35320554d068