Respect prefers-reduced-motion in carousel scroll buttons (#1537)
Adjust the carousel scroll behavior to respect the user's
prefers-reduced-motion setting, providing an instant scroll for users
who have indicated a preference for reduced motion.

Note that this does not solve the scroll-snapping: we need to figure out
a way to scroll instantly when `prefers-reduced-motion` is enabled. That
might require some work. So starting with this as a small improvement.